api-integration-demo/requirements.txt
requests>=2.31.0
httpx>=0.25.0
flask>=3.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
